refactor: oneOf/anyOf w/ discriminator common code
